refactor:refactor readability to be server function instead of endpoint for ease
