+++ src/pages/api/fetchSummary.ts
@@ -10,6 +10,7 @@ import {
 } from "~/types";
 import { openAIRequest } from "~/api-functions/open-ai-request";
 import os from "os";
+import readability from "~/api-functions/readability";
 /**
  * Make a POST request:
  * {
@@ -42,7 +43,7 @@ export default async function handler(
       code: 405,
     });
   const { url, type, text, wordLimit } = body;
-  const HOST_URL = req.headers.host;
+
   if (type === "text" && (!text || text.length <= 0))
     return res.status(400).json({
       message: "Please pass text value for this type of request",
@@ -62,7 +63,7 @@ export default async function handler(
     if (type === "text" && typeof text === "string")
       openAiResponse = await callWithText(text, wordLimit, type);
     if (type === "song" || type === "article")
-      openAiResponse = await callWithUrl(url, wordLimit, type, HOST_URL);
+      openAiResponse = await callWithUrl(url, wordLimit, type);
     return res.status(200).json({ openAiResponse });
   } catch (error) {
     res.status(500).json({ message: "Request errored out", code: 500 });
@@ -84,57 +85,34 @@ async function callWithText(
     throw error;
   }
 }
-async function callWithUrl(
-  url: string,
-  wordLimit: number,
-  type: ContentType,
-  host?: string
-) {
+async function callWithUrl(url: string, wordLimit: number, type: ContentType) {
   try {
     const CHUNK_LENGTH = 500;
     // await fetchRetry(`/readability?url_resource=${url}`, 100, 3)
-    const innerResponse = await fetch(
-      `${host}/api/readability?url_resource="${url}"&chunk_length=${CHUNK_LENGTH}`
-    );
+    const json = await readability(url, CHUNK_LENGTH);
     // if res is good, process in openAPI
-    if (innerResponse.ok) {
-      const json = await innerResponse.json();
-      const body = {
-        type,
-        chunkedTextContent: json.chunkedTextContent ?? [""],
-        wordLimit,
-      };
-      const response = await openAIRequest(body);
 
-      return {
-        ...response,
-        byline: json.byline,
-        title: json.title,
-        dir: json.dir,
-        url: json.url,
-      };
-    }
+    if (!json) throw Error("Readability failed.");
+    const body = {
+      type,
+      chunkedTextContent: json.chunkedTextContent ?? [""],
+      wordLimit,
+    };
+    const response = await openAIRequest(body);
+
+    return {
+      ...response,
+      byline: json.byline,
+      title: json.title,
+      dir: json.dir,
+      url: json.url,
+    };
   } catch (error) {
     console.log(error);
     throw error;
   }
-  throw new Error("Something went wrong");
 }
 
-async function wait(delay: number) {
-  return new Promise((resolve) => setTimeout(resolve, delay));
-}
-
-// If we have failures with the URL parsing, use this
-function fetchRetry(
-  url: string,
-  delay: number,
-  tries: number
-): Promise<Response | Error> {
-  function onError(error: Error) {
-    let triesLeft = tries - 1;
-    if (!triesLeft) throw error;
-    return wait(delay).then(() => fetchRetry(url, delay, triesLeft));
-  }
-  return fetch(url).catch((error: Error) => onError(error));
-}
+export const config = {
+  type: "experimental-background",
+};
