--- src/query/fetch-article-data.ts
@@ -1,4 +1,0 @@
-export const fetchArticleData = async (url: string) => {
-  const res = await fetch(`/api/readability?url_resource="${url}"`);
-  return await res.json();
-};
