+++ src/utils/generatePrompt.ts
